docs: improve comments and snapshot version handling
docs: Updated imports for changed file name from useReloadHandler to useSnapshotState
